Run a test comparison with current ubjson state
The sample data currently doesn't contain numbers.

Results are of the order of::

    2021-09-01 05:38:33,477 INFO number of records: 500, table size 149 MB
    ...
    2021-09-01 05:47:10,658 INFO best for bytea: 0.137608 sec
    2021-09-01 05:47:10,658 INFO best for jsonb-disk: 1.079823 sec
    2021-09-01 05:47:10,658 INFO best for ubjson: 1.251518 sec
    2021-09-01 05:47:10,658 INFO best for unparsed: 2.167307 sec
    2021-09-01 05:47:10,658 INFO best for orjson: 3.796592 sec
    2021-09-01 05:47:10,658 INFO best for jsonb: 5.698109 sec

So the ubjson loading is only <10% slower than parsing the varlena
stream, and 3x faster than the faster Python json parser.

Note: "unparsed" means that Postgres converts varlena to text but Python
doesn't parse it. It seems to mean that the overhead (the difference
between "unparsed" and "bytea") is mostly on the PostgreSQL side.